Fix editor installation issue
Installation of the editor has been failing for an unknown period
of time with the following errors:

```
renderer › The function requires a "fs" parameter but none was provided.
renderer › The function requires a "http" parameter but none was provided.
```

This was raised in [this issue](https://github.com/ministryofjustice/fb-editor-console-electron/issues/207).

Despite those parameters being passed in the cloning still fails. After
some investigation I decided to just swap it out for something simpler.