Name the Mac Documents folder correctly
Thankfully Apple does not allow a `documents` (with a lower case d)
directory to be created in the users home directory as `Documents`
already exists. It does also seem to correctly alias one to the other.

However the directory is actually with an uppercase D so let's just use
that.